otp_views: Move OTP-related views from `www/itou_staff_views/` to `www/otp_views/`
These views are currently used by staff users only, but they are going
to be used by pro users as well.
