+++ itou/templates/layout/_header_user_dropdown_menu.html
@@ -33,7 +33,7 @@
     {% endif %}
     {% if user.is_staff %}
         <li>
-            <a class="dropdown-item text-primary" href="{% url 'itou_staff_views:otp_devices' %}">Configuration OTP</a>
+            <a class="dropdown-item text-primary" href="{% url 'otp_views:otp_devices' %}">Configuration OTP</a>
         </li>
     {% endif %}
     <li class="dropdown-divider"></li>
