+++ itou/templates/otp_views/otp_confirm_device.html
